Instrument.js refactor (#721)
* fix legacy edu stuff

* pricing table semantic enhancement

* WIP

* analytics split from instrument

* fix indent CSS

* fixing more bugs

* more instrument.js refactor

* remove redundant linksPopulated Events

* less promisy refactor

* move martechLoaded promise to window

---------
